Post CRUD and API responses

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Http\Resources\PostResource;
 use App\Models\Post;
+use App\Models\PostImage;
+use App\Models\Tag;
 use Illuminate\Http\Request;
 use Validator;
 
@@ -23,12 +25,14 @@ class PostController extends BaseController
     public function store(Request $request)
     {
         try {
-            $input = $request->only(['user_id', 'group_id', 'content', 'linkable_id', 'linkable_type']);
+            $input = $request->only(['group_id', 'content', 'linkable_id', 'linkable_type']);
+            $input['user_id'] = auth()->user()->id;
 
             $validation = Validator::make($input,
                 [
                     'user_id' => 'required',
                     'content' => 'required',
+                    'images' => 'image|mimes:jpeg,png,jpg'
                 ]);
 
             if($validation->fails()){
@@ -38,7 +42,11 @@ class PostController extends BaseController
             $post = Post::create($input);
 
             if ($request->has('tags')){
-                $post->tags()->sync($request->get('tags'));
+                $post->tags()->attach(Tag::retrieveTagIds($request->get('tags')));
+            }
+
+            if ($request->hasFile('images')){
+                Post::attachImages($request->file('images'), $post->id);
             }
 
             return $this->sendResponse(new PostResource($post), 'Post created.');
@@ -57,7 +65,13 @@ class PostController extends BaseController
     public function show($post)
     {
         try {
-            return $this->sendResponse(new PostResource(Post::findorFail($post)), 'Post retrieved.');
+            $post = Post::find($post);
+
+            if ($post){
+                return $this->sendResponse(new PostResource($post), 'Post retrieved.');
+            } else {
+                return $this->sendError('No matching post found.');
+            }
 
         } catch (\Throwable $th) {
             return response()->json([
@@ -72,12 +86,15 @@ class PostController extends BaseController
      */
     public function update(Request $request, Post $post)
     {
+        if (auth()->user()->id != $post->user_id){
+            return $this->sendError('You are not the owner of the requested resource.', 'Unauthenticated.');
+        }
+
         try {
-            $input = $request->only(['user_id', 'group_id', 'content', 'linkable_id', 'linkable_type']);
+            $input = $request->only(['content', 'linkable_id', 'linkable_type']);
 
             $validation = Validator::make($input,
                 [
-                    'user_id' => 'required',
                     'content' => 'required',
                 ]);
 
@@ -88,7 +105,7 @@ class PostController extends BaseController
             $post->update($input);
 
             if ($request->has('tags')){
-                $post->tags()->sync($request->get('tags'));
+                $post->tags()->attach(Tag::retrieveTagIds($request->get('tags')));
             }
 
             return $this->sendResponse(new PostResource($post), 'Post created.');
@@ -104,8 +121,17 @@ class PostController extends BaseController
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Post $post)
+    public function destroy($post)
     {
+        $post = Post::find($post);
+        if (!$post){
+            return $this->sendError('No matching post found.');
+        }
+
+        if (auth()->user()->id != $post->user_id){
+            return $this->sendError('You are not the owner of the requested resource.', 'Unauthenticated.');
+        }
+
         $post->tags()->detach();
         $post->likes()->detach();
         $post->comments()->delete();
